% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{.extrude_walk}
\alias{.extrude_walk}
\title{Extrude a random walk.}
\usage{
.extrude_walk(walk, z = 1)
}
\arguments{
\item{walk}{a random walk.}

\item{z}{extrusion value.}
}
\value{
a data frame.
}
\description{
Extrude a random walk.
}
