% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_depth.R
\name{.identify_depth}
\alias{.identify_depth}
\title{Identify depth for plotting purposes.}
\usage{
.identify_depth(walk)
}
\arguments{
\item{walk}{an extruded random walk.}
}
\value{
a data frame.
}
\description{
Identify depth for plotting purposes.
}
