% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background.R
\name{generate_background_pattern}
\alias{generate_background_pattern}
\title{Generate a background pattern.}
\usage{
generate_background_pattern(res = 0.05)
}
\arguments{
\item{res}{resolution of the pattern.}
}
\value{
a data frame of points.
}
\description{
Generate a background pattern.
}
