% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_walk.R
\name{generate_steps}
\alias{generate_steps}
\title{Generate random steps into four wind directions.}
\usage{
generate_steps(steps)
}
\arguments{
\item{steps}{number of steps in the walk.}
}
\value{
a character vector.
}
\description{
Generate random steps into four wind directions.
}
