% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_walk.R
\name{generate_walk}
\alias{generate_walk}
\title{Generate random steps into four wind directions.}
\usage{
generate_walk(n_segments)
}
\arguments{
\item{n_segments}{number of steps in the walk.}
}
\value{
a data frame.
}
\description{
Generate random steps into four wind directions.
}
