% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background.R
\name{get_bbox}
\alias{get_bbox}
\title{Get bounding box of an object.}
\usage{
get_bbox(df, border)
}
\arguments{
\item{df}{a data frame with an x and y column.}

\item{border}{border around the bounding box.}
}
\value{
a data frame.
}
\description{
Get bounding box of an object.
}
