% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_random_walk}
\alias{plot_random_walk}
\title{Generate a ggplot2 object from random walk coordinates.}
\usage{
plot_random_walk(walk, pattern, plot_colors)
}
\arguments{
\item{walk}{a walk as generated by \code{generate_walk()}}

\item{pattern}{a background pattern with x, y and color column}

\item{plot_colors}{colors for the plot}
}
\value{
a ggplot2 object.
}
\description{
Generate a ggplot2 object from random walk coordinates.
}
