% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_walk.R
\name{plot_random_walk_from_coords}
\alias{plot_random_walk_from_coords}
\title{Generate a ggplot2 object from random walk coordinates.}
\usage{
plot_random_walk_from_coords(coords)
}
\arguments{
\item{coords}{a data frame of points.}
}
\value{
a ggplot2 object.
}
\description{
Generate a ggplot2 object from random walk coordinates.
}
