% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_walk.R
\name{random_walk_coords}
\alias{random_walk_coords}
\title{Convert a random walk to coordinates.}
\usage{
random_walk_coords(random_walk)
}
\arguments{
\item{random_walk}{a data frame as generated by \code{generate_steps()}.}
}
\value{
a data frame of points.
}
\description{
Convert a random walk to coordinates.
}
