% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background.R
\name{generate_background_pattern}
\alias{generate_background_pattern}
\title{Generate a background pattern.}
\usage{
generate_background_pattern(walk, res = 0.05, border = 1)
}
\arguments{
\item{walk}{a walk.}

\item{res}{resolution of the pattern.}

\item{border}{spacing around walk.}
}
\value{
a data frame of points.
}
\description{
Generate a background pattern.
}
