% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_walk.R
\name{generate_random_walk}
\alias{generate_random_walk}
\title{Generate a random walk and return it as a ggplot2 object.}
\usage{
generate_random_walk(steps, projection = NULL)
}
\arguments{
\item{steps}{number of steps in the walk.}

\item{projection}{projection, supported: "cabinet"}
}
\value{
a ggplot2 object.
}
\description{
Generate a random walk and return it as a ggplot2 object.
}
