% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_walk.R
\name{plot_random_walk}
\alias{plot_random_walk}
\title{Generate a ggplot2 object from random walk coordinates.}
\usage{
plot_random_walk(walk, alpha = 63.4, z = 1, levels = 3,
  spacing = 0)
}
\arguments{
\item{walk}{a walk as generated by \code{generate_walk()}}

\item{alpha}{reprojection angle axis (in degrees).}

\item{z}{extrusion value.}

\item{levels}{number of levels of the walk.}

\item{spacing}{spacing between levels of the walk.}
}
\value{
a ggplot2 object.
}
\description{
Generate a ggplot2 object from random walk coordinates.
}
